requirements.txt
langchain
chromadb
sentence-transformers
groq
streamlit
python-dotenv
PyPDF2
